+++ rt/rs/extensions/rx/pom.xml
@@ -58,6 +58,5 @@
             <artifactId>junit</artifactId>
             <scope>test</scope>
         </dependency>
-    
     </dependencies>
 </project>
