+++ systests/jaxrs/src/test/java/org/apache/cxf/systest/jaxrs/reactor/HelloWorldBean.java
@@ -16,7 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.cxf.extensions.reactor.test;
+
+package org.apache.cxf.systest.jaxrs.reactor;
 
 public class HelloWorldBean {
     private String greeting;
